+++ packages/core/strapi/lib/middlewares/ip.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const ip = require('koa-ip');
+
+module.exports = options => ip(options);
