--- packages/core/strapi/lib/middlewares/ip/defaults.json
@@ -1,7 +1,0 @@
-{
-  "ip": {
-    "enabled": false,
-    "whiteList": [],
-    "blackList": []
-  }
-}
