--- packages/core/strapi/lib/middlewares/ip/index.js
@@ -1,25 +1,0 @@
-'use strict';
-
-const ip = require('koa-ip');
-/**
- * IP filter hook
- */
-
-module.exports = strapi => {
-  return {
-    /**
-     * Initialize the hook
-     */
-
-    initialize() {
-      const { whiteList, blackList } = strapi.config.middleware.settings.ip;
-
-      strapi.server.use(
-        ip({
-          whitelist: whiteList,
-          blacklist: blackList,
-        })
-      );
-    },
-  };
-};
