+++ packages/core/strapi/lib/middlewares/logger.js
@@ -1,0 +1,25 @@
+'use strict';
+const chalk = require('chalk');
+
+const codeToColor = code => {
+  return code >= 500
+    ? chalk.red(code)
+    : code >= 400
+    ? chalk.yellow(code)
+    : code >= 300
+    ? chalk.cyan(code)
+    : code >= 200
+    ? chalk.green(code)
+    : code;
+};
+
+// TODO: inject strapi
+module.exports = () => {
+  return async (ctx, next) => {
+    const start = Date.now();
+    await next();
+    const delta = Math.ceil(Date.now() - start);
+
+    strapi.log.http(`${ctx.method} ${ctx.url} (${delta} ms) ${codeToColor(ctx.status)}`);
+  };
+};
