--- packages/core/strapi/lib/middlewares/logger/defaults.json
@@ -1,5 +1,0 @@
-{
-  "logger": {
-    "enabled": true
-  }
-}
