--- packages/core/strapi/lib/middlewares/poweredBy/defaults.json
@@ -1,5 +1,0 @@
-{
-  "poweredBy": {
-    "enabled": true
-  }
-}
