--- packages/core/strapi/lib/middlewares/poweredBy/index.js
@@ -1,16 +1,0 @@
-'use strict';
-
-module.exports = strapi => {
-  return {
-    initialize() {
-      strapi.server.use(async (ctx, next) => {
-        await next();
-
-        ctx.set(
-          'X-Powered-By',
-          strapi.config.get('middleware.settings.poweredBy.value', 'Strapi <strapi.io>')
-        );
-      });
-    },
-  };
-};
